requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.6
djangorestframework==3.7.1
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
static3==0.7.0
urllib3==1.22
psycopg2==2.7.3.2
